requirements.txt
# requirements.txt

fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
solana>=0.34.0          # python-solana client
solders>=0.21.0         # faster keypair/transaction handling
base58>=2.1.0

